Add another rating category in the CSS
This will allow styling of the rating color button on three levels
instead of two. It does make sense to potentially add either more levels
or have a gradient, based on the rating value.

The color selected for 'low' is completely arbitrary and in general, the
allocation of the colors to the rating category is arbitrary. Might make
sense to change them so that they match the theme of the site.