Include Season 1-15: Progress
Justification for inclusion:

As an engineer myself it's quintessential to teach the viewers the
difference between a warp matrix flux capacitor and a self-sealing
stem bolt.

Nobody really knows what self-sealing stem bolts are, what they are
for, what they do, or why anyone would want them. However, the
Noh-Jay consortium will gladly trade them against seven tessipates
of land on Bajor. In any case, they are better than yamok sauce.

Plus, the aliquot sum of 15 is 9, which is perfect for DS-9.